requirements.txt
flask==3.0.2
flask-socketio==5.3.6
pandas==2.2.1
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.2.3
selenium==4.18.1
urllib3<2.0.0
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
undetected-chromedriver>=3.5.5
requests==2.31.0
Flask-Assets==2.1.0
libsass==0.22.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
alembic>=1.13.1
rapidfuzz==3.13.0
flask_login==0.6.3
sendgrid==6.11.0
icalendar==5.0.11
pytz==2024.1
